Add pontoon.cmd and console_scripts to setup.py
console_scripts expects python files with function names to call, so
scripts are moved to a new sub-module pontoon.cmd. Not sure if this is
the appropriate name though.
ALso, cmd python filenames are used in underscore format in setup.py,
actual files will be renamed next.

Script files entry is not needed anymore, as console_scripts replace
them.